separate prameter_study into parameter_study and parameter_combinator (.h and .cpp both); add n_cycles to two_sinusoids
